+++ bank-feeds/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [csharp v2.3.0] bank-feeds
 ### Releases
 - [NuGet v2.3.0] https://www.nuget.org/packages/Codat.BankFeeds/2.3.0 - bank-feeds
+
+## 2023-10-30 01:27:21
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.3.1] bank-feeds
+### Releases
+- [NuGet v2.3.1] https://www.nuget.org/packages/Codat.BankFeeds/2.3.1 - bank-feeds